requirements.txt
streamlit==1.22.0
openai==0.27.8
requests==2.31.0
python-docx==0.8.11